Made modular and fit MVVM
