+++ SPAR_iOS/SPAR/SPAR/View/Device/ProcessDetailPage.swift
